requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
redis==5.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
PyJWT==2.8.0
passlib[bcrypt]
bcrypt==4.1.1
nest-asyncio==1.5.8
python-multipart==0.0.22 
bcrypt